refactor: wathced key 상수화 및 searchModal instanse화
