+++ cypress.json
@@ -1,3 +1,4 @@
 {
-    "baseUrl":"http://localhost:9001"
+    "baseUrl":"http://localhost:9001",
+    "video":false
 }
