+++ src/js/catcher/index.js
@@ -1,0 +1,8 @@
+import MainPage from './mainPage';
+import SearchModal from './searchModal';
+
+export default function initEventCatcher() {
+  const mainPage = new MainPage();
+  mainPage.init();
+  const searchModal = new SearchModal();
+}
