+++ src/js/constant/index.js
@@ -1,5 +1,7 @@
 export const LOCALSTORAGE_KEY_SAVE = 'save';
 
+export const LOCALSTORAGE_KEY_WATCHED = 'watched';
+
 export const ERROR_403 = '403 Error';
 
 export const REQUEST_VIDEO_QUANTITY = 10;
