+++ src/js/presenter/templates.js
